requirements.txt
einops==0.4.1
joblib==1.1.0
numpy==1.21.2
opencv-python==4.4.0.46
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@0d53acbcf5123d4c88040fe68fbb9805fc5b2fb9
Pillow==8.4.0
timm==0.3.2
torch==1.12.0+cu116
torchvision==0.13.0+cu116
tqdm==4.63.0